refactor: fix issue of response necver being sent and some more renaming
