+++ src/main/ResponseManager.ts
